optimize bmi270 driver
